test(direct-control): extract operation request tests
Move package-owned unit-operation and production-choice coverage into focused direct-control package tests with local fake tuner fixtures.

Record the operation package-test slice as DRA-owned because background thread creation was unavailable. Keep operation source extraction pending for remaining unit-target and chooser-closeout package tests.

Validation: bun run --cwd packages/civ7-direct-control test -- unit-operation.test.ts production-choice.test.ts; bun run --cwd packages/civ7-direct-control test -- unit-operation.test.ts production-choice.test.ts direct-control.test.ts; bun run --cwd packages/civ7-direct-control test; bun run --cwd packages/civ7-direct-control check; bun run --cwd packages/civ7-direct-control build; bun run openspec -- validate civ7-support-direct-control-modularization --strict; git diff --check.
