+++ openspec/changes/civ7-support-direct-control-modularization/workstream/direct-control-atom-corpus.md
@@ -15,7 +15,7 @@ extraction starts with package-owned atom tests, not source moves.
 | Runtime command source builders | `index.ts` lines 3909+ and 4714-4847: command builders and embedded JS helpers | `src/runtime-sources/*` and atom-local source modules | Direct package tests often match function names; CLI consumers invoke package wrappers | Exact string/source behavior equivalence tests before moves | runtime proof required for embedded JS behavior changes | Planned; high-risk source surface. |
 | Notification HUD/materialization | Play types around 913-1000, wrapper around 2925, builder around 4714, source around 5119+ | `src/play/notifications/{view,details,decision-hints}.ts` | `test/play-notification-view.test.ts` now owns package-level `getCiv7PlayNotificationView` materialization and decision-hint coverage. CLI owns the broader scenario matrix in notification HUD, queue, and priorities suites. | Before source moves, preserve package API shape and keep scenario expansion out of a broad shared notification catalog unless named consumers require it. | runtime proof for live HUD/source behavior changes | Package proof completed; source extraction still pending. |
 | Notification dismissal/verification | Types around 1005-1045, wrappers around 2935/2946, source around 7491, polling/identity verification around 10472-10534 | `src/play/notifications/{dismissal,verification}.ts` | `test/notification-dismissal.test.ts` now owns guarded read/send, verified dismissal, engine-front train absence, engine-front dismissed, none-blocker panel dismiss, and expired non-dismissible none-blocker coverage. CLI owns bulk-dismiss policy and stale nonblocking command behavior. | Before source moves, preserve verification semantics and App UI route proof. Do not move CLI bulk-dismiss candidate selection into direct-control. | runtime proof for timing/routes/identity behavior changes | Package proof completed; source extraction still pending. |
-| Operation validation/send/postconditions | Operation types around 1190-1341, wrappers around 2996-3130, builders 4722/4729, router around 7753, classifiers 10660-11224; diplomacy/narrative wrappers around 3194/3246 | `src/play/operations/{families,validate-request,postconditions,production-choice,chooser-closeouts}.ts` | Package covers unit operation, production choice, unit target basic/no-op, technology closeout, and culture closeout. CLI has delayed/path-shortfall unit-target cases and diplomacy/narrative coverage. | Add focused tests: `unit-operation`, `production-choice`, `unit-target-action`, `technology-choice-closeout`, `culture-choice-closeout`, `diplomacy-response`, and `narrative-choice`; preserve approval-first semantics and no-repeat-after-unverified semantics. | runtime proof for mutation/postcondition behavior changes | Planned after CLI command surfaces stabilize; diplomacy/narrative package proof missing. |
+| Operation validation/send/postconditions | Operation types around 1190-1341, wrappers around 2996-3130, builders 4722/4729, router around 7753, classifiers 10660-11224; diplomacy/narrative wrappers around 3194/3246 | `src/play/operations/{families,validate-request,postconditions,production-choice,chooser-closeouts}.ts` | `test/unit-operation.test.ts` now owns validator-first approved unit-operation send/postcondition coverage. `test/production-choice.test.ts` now owns the official App UI production path and production-choice-cleared postcondition. The broad suite still owns unit-target basic/no-op, technology closeout, and culture closeout until those focused files move. CLI has delayed/path-shortfall unit-target cases and diplomacy/narrative coverage. | Add remaining focused tests: `unit-target-action`, `technology-choice-closeout`, `culture-choice-closeout`, `diplomacy-response`, and `narrative-choice`; preserve approval-first semantics and no-repeat-after-unverified semantics. | runtime proof for mutation/postcondition behavior changes | Unit-operation and production-choice package proof completed; remaining operation tests and source extraction still pending. |
 | Ready unit/city/move preview | Ready-unit types 1392-1449, move preview 1451-1482, ready-city 1484-1549, wrappers 3313/3328/3344, sources 9094/9343/9481 | `src/play/ready/{unit,city,move-preview}.ts` | Package now covers `getCiv7ReadyUnitView`, `getCiv7ReadyCityView`, and `getCiv7UnitMovePreview` in focused files. Ready-view tests own read-only/no-send assertions and local minimal fixtures; unit-move-preview owns read shape, command routing, bounds rejection, and conservative relationship policy. CLI covers compact envelopes. | Before source moves, preserve public API shape and split source by ready unit, ready city, and move preview owners without creating a broad ready/common catalog. | runtime proof for live read/source behavior changes | Ready package proof completed; source extraction still pending. |
 | Tactical/progression/read lenses | Unit target source 8041, settlement 8258, target candidates 8349, battlefield 8637, destination 8927, traditions 4864, progress dashboard 4975 | `src/play/tactical/*` and `src/play/progression/*` | CLI owns settlement, progression, tactical, and priorities consumer coverage; package lacks direct tests for settlement, target candidates, traditions, progress dashboard, battlefield, and destination reads. | Add `test/settlement-recommendations.test.ts`, `test/traditions-view.test.ts`, `test/progress-dashboard.test.ts`, `test/target-candidates.test.ts`, `test/battlefield-scan.test.ts`, and `test/destination-analysis.test.ts`; assert conservative relationship labels. | read-only live proof only when claiming runtime behavior | Planned; relationship invariant protected. |
 | Map/visibility/gameinfo reads | `index.ts` wrappers around 2358-2469 plus bounded map/grid source regions | `src/play/map/{summary,plot,grid,visibility,gameinfo}.ts` | Package covers map summary, plot snapshot, bounded grid cap, visibility, reveal, and GameInfo rows in broad suite. | Add `test/map-and-visibility.test.ts`; preserve hiddenInfoPolicy and local bounds caps before source moves. | runtime proof only for reveal/mutation behavior changes | Newly added row from package-test boundary report. |
